fix: move scripts to dedicated package
the goal is to remove a circular dep
