+++ packages/cmf-webpack-plugin/package.json
@@ -28,12 +28,10 @@
   },
   "homepage": "https://github.com/Talend/ui/blob/master/packages/cmf/README.md",
   "dependencies": {
+    "@talend/scripts-cmf": "^1.0.0",
     "lodash": "^4.17.21"
   },
   "devDependencies": {},
-  "peerDependencies": {
-    "@talend/react-cmf": ">=7.0.2"
-  },
   "publishConfig": {
     "access": "public"
   },
