+++ packages/cmf/package.json
@@ -36,6 +36,7 @@
     "cmf-settings": "./scripts/cmf-settings.js"
   },
   "dependencies": {
+    "@talend/scripts-cmf": "^1.0.0",
     "deepmerge": "^1.5.2",
     "hoist-non-react-statics": "^3.3.2",
     "immutable": "^3.8.2",
