+++ tools/scripts-cmf/cmf-settings.i18n.test.js
