+++ tools/scripts-cmf/cmf-settings.utils.js
