+++ tools/scripts-cmf/cmf-settings.utils.test.js
