+++ tools/scripts-cmf/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@talend/scripts-cmf",
+  "description": "@talend/react-cmf scripts to manage CMF settings",
+  "version": "1.0.0",
+  "main": "cmf-settings.js",
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Talend/ui.git"
+  }
+}
